split Constants.java into multiple files
