+++ src/main/java/frc/robot/commands/TeleopSwerve.java
@@ -5,7 +5,9 @@ import java.util.function.DoubleSupplier;
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.Constants;
+import frc.robot.constants.AutoConstants;
+import frc.robot.constants.Constants;
+import frc.robot.constants.swerve.SwerveConstants;
 import frc.robot.subsystems.drive.DriveTrainSubsystem;
 
 public class TeleopSwerve extends CommandBase {
@@ -28,11 +30,11 @@ public class TeleopSwerve extends CommandBase {
     public void execute() {
         Translation2d driveTranslation = new Translation2d(
                 MathUtil.applyDeadband(forwardSupplier.getAsDouble(), Constants.controllerDeadband)
-                        * Constants.Swerve.maxSpeed,
+                        * SwerveConstants.maxSpeed,
                 MathUtil.applyDeadband(strafeSupplier.getAsDouble(), Constants.controllerDeadband)
-                        * Constants.Swerve.maxSpeed);
+                        * SwerveConstants.maxSpeed);
         double rotation = MathUtil.applyDeadband(this.rotationSupplier.getAsDouble(), Constants.controllerDeadband)
-                * Constants.Swerve.maxAngularVelocity;
+                * AutoConstants.MaxAngularSpeedRadiansPerSecond;
 
         this.driveTrain.drive(driveTranslation, rotation, false);
     }
