+++ src/main/java/frc/robot/constants/Constants.java
@@ -1,0 +1,11 @@
+
+package frc.robot.constants;
+
+/**
+ * Various constants used throughout the program are defined here. They're
+ * defined here instead of elsewhere just for ease of changing them.
+ */
+public final class Constants {
+    public static final boolean simReplay = false;
+    public static final double controllerDeadband = 0.1;
+}
