+++ src/main/java/frc/robot/subsystems/drive/DriveTrainSubsystem.java
@@ -16,8 +16,8 @@ import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardLayout;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants;
 import frc.robot.SimManager;
+import frc.robot.constants.swerve.SwerveConstants;
 import frc.robot.sensors.imu.IMU;
 
 /**
@@ -46,7 +46,7 @@ public class DriveTrainSubsystem extends SubsystemBase {
         Timer.delay(1);
         this.resetModules();
 
-        this.swerveOdometry = new SwerveDriveOdometry(Constants.Swerve.swerveKinematics, imu.getHeading(), this.getModulePositions());
+        this.swerveOdometry = new SwerveDriveOdometry(SwerveConstants.swerveKinematics, imu.getHeading(), this.getModulePositions());
         if (RobotBase.isSimulation()) {
             SimManager.getInstance().registerDriveTrain(this::getPose, this::getSpeed);
         }
@@ -63,7 +63,7 @@ public class DriveTrainSubsystem extends SubsystemBase {
 
     public void drive(Translation2d translation, double rotation, boolean fieldRelative) {
         SwerveModuleState[] swerveModuleStates =
-            Constants.Swerve.swerveKinematics.toSwerveModuleStates(
+            SwerveConstants.swerveKinematics.toSwerveModuleStates(
                 fieldRelative ? ChassisSpeeds.fromFieldRelativeSpeeds(
                                     translation.getX(), 
                                     translation.getY(), 
@@ -75,7 +75,7 @@ public class DriveTrainSubsystem extends SubsystemBase {
                                     translation.getY(), 
                                     rotation)
                                 );
-        SwerveDriveKinematics.desaturateWheelSpeeds(swerveModuleStates, Constants.Swerve.maxSpeed);
+        SwerveDriveKinematics.desaturateWheelSpeeds(swerveModuleStates, SwerveConstants.maxSpeed);
 
         for (int i = 0; i < this.swerveModules.length; i++) {
             Logger.getInstance().recordOutput("/Swerve/moduleState" + i, swerveModuleStates[i]);
@@ -87,7 +87,7 @@ public class DriveTrainSubsystem extends SubsystemBase {
     }
     
     public ChassisSpeeds getSpeed() {
-        return Constants.Swerve.swerveKinematics.toChassisSpeeds(this.getModuleStates());
+        return SwerveConstants.swerveKinematics.toChassisSpeeds(this.getModuleStates());
     }
     
     public Pose2d getPose() {
